requirements.txt
python-telegram-bot==13.4.1
requests==2.28.0
PyYAML==6.0